requirements.txt
# =================================
# SmartGroupBot Dependencies
# =================================

# Core Telegram Bot Framework
python-telegram-bot==20.7

# AI & Machine Learning
openai==1.6.1

# Database & ORM
sqlalchemy==2.0.25
aiosqlite==0.19.0
psycopg2-binary==2.9.9

# Web Framework & HTTP
flask==3.0.0
requests==2.31.0

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging & Utilities
loguru==0.7.2
python-dateutil==2.8.2

# Note: asyncio==3.4.3 is built into Python 3.7+, removed redundant dependency 